Adapter audit - part 1.
