--- packages/adapters/src/Alert/Alert.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import { Alert as ChakraAlert, AlertIcon } from '@chakra-ui/core';
-
-import type { AlertProps } from './types';
-
-const Alert: React.FC<AlertProps> = ({ description, ...props }) => (
-	<ChakraAlert {...props}>
-		<AlertIcon />
-		{description}
-	</ChakraAlert>
-);
-
-export default Alert;
