--- packages/adapters/src/Alert/index.ts
@@ -1,3 +1,0 @@
-export { default as Alert } from './Alert';
-
-export * from './types';
