--- packages/adapters/src/Alert/types.ts
@@ -1,6 +1,0 @@
-import type { AlertProps as ChakraAlertProps, IconProps } from '@chakra-ui/core';
-
-export interface AlertProps extends ChakraAlertProps {
-	description: string;
-	iconProps?: IconProps;
-}
