--- packages/adapters/src/Avatar/index.ts
@@ -1,3 +1,0 @@
-export { default as Avatar } from './Avatar';
-
-export * from './types';
