--- packages/adapters/src/Avatar/types.ts
@@ -1,6 +1,0 @@
-import type { BoxProps, IAvatar } from '@chakra-ui/core';
-
-export interface AvatarProps extends IAvatar {
-	badgeProps?: BoxProps;
-	userName?: string;
-}
