+++ packages/adapters/src/Badge/Badge.tsx
@@ -3,6 +3,6 @@ import { Badge as ChakraBadge } from '@chakra-ui/core';
 
 import type { BadgeProps } from './types';
 
-const Badge: React.FC<BadgeProps> = ({ children, ...props }) => <ChakraBadge {...props}>{children}</ChakraBadge>;
+const Badge: React.FC<BadgeProps> = ({ children, style }) => <ChakraBadge style={style}>{children}</ChakraBadge>;
 
 export default Badge;
