+++ packages/adapters/src/Badge/types.ts
@@ -1,3 +1,3 @@
 import type { BadgeProps as ChakraBadgeProps } from '@chakra-ui/core';
 
-export interface BadgeProps extends ChakraBadgeProps {}
+export interface BadgeProps extends Pick<ChakraBadgeProps, 'className' | 'style'> {}
