+++ packages/adapters/src/Banner/types.ts
@@ -1,6 +1,6 @@
 import type { AlertProps, IconProps } from '@chakra-ui/core';
 
-export interface BannerProps extends AlertProps {
+export interface BannerProps extends Pick<AlertProps, 'className' | 'status' | 'variant'> {
 	description?: string;
 	iconProps?: IconProps;
 	title: string;
