+++ packages/components/src/Banner/Banner.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 
-import { Banner as BannerAdapter, BannerProps } from '@eventespresso/adapters';
+import { Banner as BannerAdapter } from '@eventespresso/adapters';
+import type { BannerProps } from '@eventespresso/adapters';
 
 import './style.scss';
