+++ packages/components/src/EmptyState/types.ts
@@ -1,0 +1,6 @@
+export interface EmptyStateProps {
+	children?: React.ReactNode;
+	className?: string;
+	description: string;
+	title?: string;
+}
