+++ packages/components/src/ErrorIndicator/style.scss
@@ -1,0 +1,5 @@
+.ee-error-indicator {
+	flex-direction: column;
+	justify-content: center;
+	text-align: center;
+}
