+++ packages/components/src/ErrorIndicator/types.ts
@@ -1,0 +1,4 @@
+export interface ErrorIndicatorProps {
+	description?: string;
+	title?: string;
+}
